refactor: move SMILESEncoderDataset to smiles_dataset.py.
